requirements.txt
# Core Framework
Flask>=3.0.0,<4.0.0
Jinja2>=3.1.0,<4.0.0
Werkzeug>=3.0.0,<4.0.0
Click>=8.1.0,<9.0.0
Waitress

# Security
Flask-WTF>=1.2.0,<2.0.0

# Database Support
# SQLite is built into Python
psycopg2-binary>=2.9.0  # PostgreSQL
pymssql>=2.2.0          # Microsoft SQL Server
# Uncomment below for other database support
# SQLAlchemy>=2.0.0,<3.0.0
# pymysql>=1.1.0          # MySQL

# Development & Testing
pytest>=7.4.0,<8.0.0
pytest-flask>=1.3.0,<2.0.0
pytest-cov>=4.1.0,<5.0.0

# Production Server (bundled for standalone binary)
waitress>=3.0.0         # Cross-platform production WSGI server

# Build
PyInstaller>=6.0.0,<7.0.0
